+++ substratevm/src/com.oracle.svm.core.genscavenge/src/com/oracle/svm/core/genscavenge/OldGeneration.java
@@ -32,45 +32,27 @@ import static jdk.graal.compiler.nodes.extended.BranchProbabilityNode.VERY_SLOW_
 import static jdk.graal.compiler.nodes.extended.BranchProbabilityNode.probability;
 
 import com.oracle.svm.core.NeverInline;
-import com.oracle.svm.core.SubstrateOptions;
-import com.oracle.svm.core.UnmanagedMemoryUtil;
 import com.oracle.svm.core.code.RuntimeCodeInfoMemory;
-import com.oracle.svm.core.config.ConfigurationValues;
+import com.oracle.svm.core.genscavenge.tenured.AllObjectsMarkingVisitor;
+import com.oracle.svm.core.genscavenge.tenured.CompactingVisitor;
+import com.oracle.svm.core.genscavenge.tenured.FixingVisitor;
+import com.oracle.svm.core.genscavenge.tenured.PlanningVisitor;
+import com.oracle.svm.core.genscavenge.tenured.RefFixingVisitor;
+import com.oracle.svm.core.genscavenge.tenured.RelocationInfo;
 import com.oracle.svm.core.graal.RuntimeCompilation;
-import com.oracle.svm.core.heap.Heap;
-import com.oracle.svm.core.heap.ObjectReferenceVisitor;
-import com.oracle.svm.core.heap.ReferenceAccess;
-import com.oracle.svm.core.heap.ReferenceInternals;
-import com.oracle.svm.core.hub.DynamicHub;
-import com.oracle.svm.core.hub.InteriorObjRefWalker;
-import com.oracle.svm.core.identityhashcode.IdentityHashCodeSupport;
-import com.oracle.svm.core.snippets.KnownIntrinsics;
-import com.oracle.svm.core.stack.JavaStackWalk;
-import com.oracle.svm.core.stack.JavaStackWalker;
-import com.oracle.svm.core.thread.VMThreads;
-import org.graalvm.compiler.word.ObjectAccess;
-import org.graalvm.compiler.word.Word;
-import org.graalvm.nativeimage.CurrentIsolate;
-import org.graalvm.nativeimage.IsolateThread;
 import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
-import org.graalvm.nativeimage.StackValue;
-import org.graalvm.nativeimage.c.function.CodePointer;
 import org.graalvm.word.Pointer;
 import org.graalvm.word.UnsignedWord;
-import org.graalvm.word.WordFactory;
 
 import com.oracle.svm.core.AlwaysInline;
 import com.oracle.svm.core.Uninterruptible;
 import com.oracle.svm.core.genscavenge.GCImpl.ChunkReleaser;
 import com.oracle.svm.core.genscavenge.remset.RememberedSet;
 import com.oracle.svm.core.heap.ObjectVisitor;
-import com.oracle.svm.core.hub.LayoutEncoding;
 import com.oracle.svm.core.log.Log;
 import com.oracle.svm.core.thread.VMOperation;
 import com.oracle.svm.core.util.VMError;
-
-import java.lang.ref.Reference;
 
 /**
  * The old generation has only one {@link Space} for existing, newly-allocated or promoted objects
@@ -82,8 +64,8 @@ public final class OldGeneration extends Generation {
 
     private final GreyObjectsWalker toGreyObjectsWalker = new GreyObjectsWalker();
     private final PlanningVisitor planningVisitor = new PlanningVisitor();
-    private final FixingVisitor fixingVisitor = new FixingVisitor();
     private final RefFixingVisitor refFixingVisitor = new RefFixingVisitor();
+    private final FixingVisitor fixingVisitor = new FixingVisitor(refFixingVisitor);
     private final CompactingVisitor compactingVisitor = new CompactingVisitor();
     private final AllObjectsMarkingVisitor allObjectsMarkingVisitor = new AllObjectsMarkingVisitor();
     private final RuntimeCodeCacheWalker runtimeCodeCacheWalker = new RuntimeCodeCacheWalker(refFixingVisitor);
@@ -321,527 +303,5 @@ public final class OldGeneration extends Generation {
         }
         RememberedSet.get().enableRememberedSetForChunk(chunk);
         return chunk;
-    }
-
-    private class AllObjectsMarkingVisitor implements ObjectVisitor {
-
-        @Override
-        public boolean visitObject(Object obj) {
-            ObjectHeaderImpl.setMarkedBit(obj);
-            return true;
-        }
-    }
-
-    private class PlanningVisitor implements ObjectVisitor {
-
-        private Pointer relocationInfoPointer;
-
-        private Pointer relocationPointer;
-
-        private AlignedHeapChunk.AlignedHeader chunk;
-
-        private UnsignedWord gapSize;
-
-        @NeverInline("Debug")
-        public boolean visitObject(Object obj) {
-            //Log.log().string("B").newline().flush();
-            Pointer objPointer = Word.objectToUntrackedPointer(obj);
-            UnsignedWord objSize = LayoutEncoding.getSizeFromObjectInGC(obj);
-            if (ObjectHeaderImpl.hasMarkedBit(obj)) {
-                // ObjectHeaderImpl.clearMarkedBit(obj); TODO: clear later
-                if (gapSize.notEqual(0)) {
-                    Log.log().string("Gap from ").zhex(objPointer.subtract(gapSize))
-                            .string(" to ").zhex(objPointer)
-                            .string(" (").unsigned(gapSize).string(" bytes)")
-                            .newline().flush();
-
-                    if (relocationInfoPointer.isNull()) {
-                        if (chunk.isNull()){
-                            Log.log().string("Nooooo! chunk=").zhex(chunk).newline().flush();
-                        }
-                        chunk.setFirstRelocationInfo(objPointer);
-                    } else {
-                        int offset = (int) objPointer.subtract(relocationInfoPointer).rawValue();
-                        RelocationInfo.writeNextPlugOffset(relocationInfoPointer, offset);
-
-                        Log.log().string("Updated relocation info at ").zhex(relocationInfoPointer)
-                                .string(": nextPlugOffset=").zhex(offset)
-                                .newline().flush();
-                    }
-                    relocationInfoPointer = objPointer;
-                    RelocationInfo.writeRelocationPointer(relocationInfoPointer, relocationPointer);
-                    RelocationInfo.writeGapSize(relocationInfoPointer, (int) gapSize.rawValue());
-                    RelocationInfo.writeNextPlugOffset(relocationInfoPointer, 0);
-
-                    Log.log().string("Wrote relocation info at ").zhex(relocationInfoPointer)
-                            .string(": relocationPointer=").zhex(relocationPointer)
-                            .string(": gapSize=").unsigned(gapSize)
-                            .string(": nextPlugOffset=").zhex(0)
-                            .newline().flush();
-
-                    gapSize = WordFactory.zero();
-                }
-                relocationPointer = relocationPointer.add(getMovedObjectSize(obj));
-            } else {
-                gapSize = gapSize.add(objSize);
-                if (gapSize.unsignedRemainder(64).equal(0)) {
-                    Log.noopLog().string("Added ").unsigned(objSize).string(" to gap size => ").unsigned(gapSize)
-                            .string(", obj=").object(obj).newline().flush();
-                }
-            }
-            return true;
-        }
-
-        public void init(AlignedHeapChunk.AlignedHeader chunk) {
-            this.chunk = chunk;
-            this.relocationInfoPointer = WordFactory.nullPointer();
-            this.relocationPointer = AlignedHeapChunk.getObjectsStart(chunk);
-            this.gapSize =  WordFactory.zero();
-        }
-
-        public void finish() {
-            if (gapSize.notEqual(0)) {
-                Pointer topPointer = HeapChunk.getTopPointer(chunk);
-                Pointer gapStart = topPointer.subtract(gapSize);
-
-                Log.log().string("Gap at chunk end from ").zhex(gapStart)
-                        .string(" to ").zhex(topPointer)
-                        .string(" (").unsigned(gapSize).string(" bytes)")
-                        .newline().flush();
-
-                HeapChunk.setTopPointer(chunk, gapStart);
-            }
-        }
-    }
-
-    private class FixingVisitor implements ObjectVisitor {
-
-        private AlignedHeapChunk.AlignedHeader chunk;
-
-        @Override
-        public boolean visitObject(Object obj) {
-
-            // fixes Target_java_lang_ref_Reference.referent
-            /*
-            DynamicHub hub = KnownIntrinsics.readHub(obj);
-            if (probability(SLOW_PATH_PROBABILITY, hub.isReferenceInstanceClass())) {
-                Reference<?> dr = (Reference<?>) obj;
-
-                Pointer objRef = ReferenceInternals.getReferentFieldAddress(dr);
-
-                Pointer p = ReferenceAccess.singleton().readObjectAsUntrackedPointer(objRef, true);
-
-                if (p.isNonNull() && !Heap.getHeap().isInImageHeap(p)) {
-                    Log.log().string("A").newline().flush();
-                    if (ObjectHeaderImpl.hasMarkedBit(p.toObject())) {
-                        Log.log().string("B").newline().flush();
-                        refFixingVisitor.visitObjectReference(objRef, true, dr);
-                    } else {
-                        Log.log().string("C").newline().flush();
-                        ReferenceInternals.setReferent(dr, null); // dead
-                        Reference<?> next = (ReferenceObjectProcessing.rememberedRefsList != null) ? ReferenceObjectProcessing.rememberedRefsList : dr;
-                        ReferenceInternals.setNextDiscovered(dr, next);
-                        ReferenceObjectProcessing.rememberedRefsList = dr;
-                    }
-                }
-            }
-            */
-
-            InteriorObjRefWalker.walkObjectInline(obj, refFixingVisitor);
-            return true;
-        }
-
-        public void setChunk(AlignedHeapChunk.AlignedHeader chunk) {
-            this.chunk = chunk;
-            refFixingVisitor.setChunk(chunk);
-        }
-    }
-
-    private class RefFixingVisitor implements ObjectReferenceVisitor {
-
-        public boolean debug = false;
-        private AlignedHeapChunk.AlignedHeader chunk;
-
-        @Override
-        public boolean visitObjectReference(Pointer objRef, boolean compressed, Object holderObject) {
-            return visitObjectReferenceInline(objRef, 0, compressed, holderObject);
-        }
-
-        @Override
-        public boolean visitObjectReferenceInline(Pointer objRef, int innerOffset, boolean compressed, Object holderObject) {
-            Pointer offsetP = ReferenceAccess.singleton().readObjectAsUntrackedPointer(objRef, compressed);
-            assert offsetP.isNonNull() || innerOffset == 0;
-
-            Pointer p = offsetP.subtract(innerOffset);
-            if (p.isNull() || HeapImpl.getHeapImpl().isInImageHeap(p)) {
-                return true;
-            }
-
-            Object obj = p.toObject();
-            if (ObjectHeaderImpl.isUnalignedObject(obj)) {
-                if (HeapImpl.getHeapImpl().isInImageHeap(holderObject)){
-                    RememberedSet.get().dirtyCardIfNecessary(holderObject, obj);
-                }
-                return true;
-            }
-
-            if (ReferenceObjectProcessing.isInOldSpace(holderObject) && !ObjectHeaderImpl.hasMarkedBit(holderObject)) {
-                Log.log().string("[ERROR: Fixing dead holder object died, obj=").object(obj)
-                        .string(", holderObject=").object(holderObject)
-                        .string("]\n").flush();
-                assert false : "Referred object died!";
-            }
-
-            if (!ObjectHeaderImpl.hasMarkedBit(obj)) {
-                Log.log().string("[ERROR: Referred object died, obj=").object(obj)
-                        .string(", holderObject=").object(holderObject)
-                        .string("]\n").flush();
-                assert false : "Referred object died!";
-            }
-
-            Pointer relocationInfo = AlignedHeapChunk.getEnclosingChunkFromObjectPointer(p).getFirstRelocationInfo();
-            if (relocationInfo.isNull() || relocationInfo.aboveThan(p)) {
-                if (HeapImpl.getHeapImpl().isInImageHeap(holderObject)){
-                    RememberedSet.get().dirtyCardIfNecessary(holderObject, obj);
-                }
-                return true;
-            }
-
-            Pointer nextRelocationInfo = RelocationInfo.getNextRelocationInfo(relocationInfo);
-            while (nextRelocationInfo.isNonNull() && nextRelocationInfo.belowOrEqual(p)) {
-                relocationInfo = nextRelocationInfo;
-                nextRelocationInfo = RelocationInfo.getNextRelocationInfo(relocationInfo);
-            }
-
-            Pointer relocationPointer = RelocationInfo.readRelocationPointer(relocationInfo);
-            Pointer newLocationOld = relocationPointer.add(p.subtract(relocationInfo));
-
-            Pointer newLocation = getRelocatedObjectPointer(p);
-            assert newLocation.equal(newLocationOld);
-
-            Object offsetObj = (innerOffset == 0) ? newLocation.toObject() : newLocation.add(innerOffset).toObject();
-            ReferenceAccess.singleton().writeObjectAt(objRef, offsetObj, compressed);
-
-            // if (ObjectHeaderImpl.hasRememberedSet(ObjectHeaderImpl.readHeaderFromObject(holderObject))) {
-            //     RememberedSet.get().dirtyCardIfNecessary(holderObject, newLocation.toObject());
-            // }
-            if (HeapImpl.getHeapImpl().isInImageHeap(holderObject)){
-                RememberedSet.get().dirtyCardIfNecessary(holderObject, newLocation.toObject());
-            }
-
-            if (debug) {
-                Log.log().string("Updated location, old=").zhex(p)
-                        .string(", new=").zhex(newLocation)
-                        .string(", diff=").signed(newLocation.subtract(p))
-                        .string(", holderObject=").object(holderObject)
-                        .string(", objRef=").zhex(objRef)
-                        .string(", innerOffset=").signed(innerOffset)
-                        .string(", compressed=").bool(compressed)
-                        .string(", relocationInfo=").zhex(relocationInfo)
-                        .string(", relocationPointer=").zhex(relocationPointer)
-                        .newline().flush();
-            }
-
-            if (!ObjectHeaderImpl.hasMarkedBit(obj)) {
-                Log.log().string("Updated location but object is not marked, old=").zhex(p)
-                        .string(", new=").zhex(newLocation)
-                        .string(", diff=").signed(newLocation.subtract(p))
-                        .string(", holderObject=").object(holderObject)
-                        .string(", objRef=").zhex(objRef)
-                        .string(", innerOffset=").signed(innerOffset)
-                        .string(", compressed=").bool(compressed)
-                        .string(", relocationInfo=").zhex(relocationInfo)
-                        .string(", relocationPointer=").zhex(relocationPointer)
-                        .newline().flush();
-            }
-
-            chunk = AlignedHeapChunk.getEnclosingChunkFromObjectPointer(p);
-            if (AlignedHeapChunk.getObjectsStart(chunk).aboveThan(newLocation) || AlignedHeapChunk.getObjectsEnd(chunk).belowOrEqual(newLocation)) {
-                Log.log().string("ERROR - Updated location, old=").zhex(p)
-                        .string(", new=").zhex(newLocation)
-                        .string(", diff=").signed(newLocation.subtract(p))
-                        .string(", holderObject=").zhex(Word.objectToUntrackedPointer(holderObject))
-                        .string(", objRef=").zhex(objRef)
-                        .string(", relocationInfo=").zhex(relocationInfo)
-                        .string(", relocationPointer=").zhex(relocationPointer)
-                        .newline().flush();
-            }
-
-            if (p.and(15).isNull()) {
-                Log.noopLog().string("Updated location, old=").zhex(p)
-                        .string(", new=").zhex(newLocation)
-                        .string(", diff=").signed(newLocation.subtract(p))
-                        .string(", holderObject=").zhex(Word.objectToUntrackedPointer(holderObject))
-                        .string(", objRef=").zhex(objRef)
-                        .string(", relocationInfo=").zhex(relocationInfo)
-                        .string(", relocationPointer=").zhex(relocationPointer)
-                        .newline().flush();
-            }
-
-            return true;
-        }
-
-        public void setChunk(AlignedHeapChunk.AlignedHeader chunk) {
-            this.chunk = chunk;
-        }
-    }
-
-
-    private class CompactingVisitor implements ObjectVisitor {
-
-        private Pointer relocationPointer;
-        private Pointer relocationInfoPointer;
-        private Pointer nextRelocationInfoPointer;
-
-        private AlignedHeapChunk.AlignedHeader chunk;
-
-        @Override
-        public boolean visitObject(Object obj) {
-            ObjectHeaderImpl.clearMarkedBit(obj); // TODO: ?
-            if (relocationInfoPointer.isNull()) {
-                return true;
-            }
-
-            Pointer objPointer = Word.objectToUntrackedPointer(obj);
-            if (objPointer.belowThan(relocationInfoPointer)) {
-                return true;
-            }
-
-            if (nextRelocationInfoPointer.isNonNull() && objPointer.aboveOrEqual(nextRelocationInfoPointer)) {
-                relocationInfoPointer = nextRelocationInfoPointer;
-                nextRelocationInfoPointer = RelocationInfo.getNextRelocationInfo(relocationInfoPointer);
-                relocationPointer = RelocationInfo.readRelocationPointer(relocationInfoPointer);
-            }
-
-            Pointer newLocation = objPointer.subtract(relocationInfoPointer).add(relocationPointer);
-
-            Log.noopLog().string("New relocation")
-                    .string(", newLocation=").zhex(newLocation)
-                    .string(", objPointer=").zhex(objPointer)
-                    .string(", relocationInfoPointer=").zhex(relocationInfoPointer)
-                    .string(", relocationPointer=").zhex(relocationPointer)
-                    .newline().flush();
-
-            UnsignedWord copySize = copyObject(obj, newLocation);
-
-            HeapChunk.setTopPointer(chunk, newLocation.add(copySize));
-
-            return true;
-        }
-
-        public void setChunk(AlignedHeapChunk.AlignedHeader chunk) {
-            this.chunk = chunk;
-            relocationInfoPointer = chunk.getFirstRelocationInfo();
-            nextRelocationInfoPointer = WordFactory.nullPointer();
-            relocationPointer = WordFactory.nullPointer();
-            if (relocationInfoPointer.isNonNull()) {
-                int offset = RelocationInfo.readNextPlugOffset(relocationInfoPointer);
-                if (offset > 0) {
-                    nextRelocationInfoPointer = relocationInfoPointer.add(offset);
-                }
-                relocationPointer = RelocationInfo.readRelocationPointer(relocationInfoPointer);
-            }
-        }
-    }
-
-    private static UnsignedWord getMovedObjectSize(Object obj) {
-        if (!ConfigurationValues.getObjectLayout().hasFixedIdentityHashField()) {
-            Word header = ObjectHeaderImpl.readHeaderFromObject(obj);
-            if (probability(SLOW_PATH_PROBABILITY, ObjectHeaderImpl.hasIdentityHashFromAddressInline(header))) {
-                return LayoutEncoding.getSizeFromObjectInlineInGC(obj, true);
-            }
-        }
-        return LayoutEncoding.getSizeFromObjectInlineInGC(obj, false);
-    }
-
-    private UnsignedWord copyObject(Object obj, Pointer dest) {
-        assert VMOperation.isGCInProgress();
-        assert ObjectHeaderImpl.isAlignedObject(obj);
-
-        // TODO: code cleanup
-
-        UnsignedWord originalSize = LayoutEncoding.getSizeFromObjectInlineInGC(obj, false);
-        UnsignedWord copySize = originalSize;
-        boolean addIdentityHashField = false;
-        if (!ConfigurationValues.getObjectLayout().hasFixedIdentityHashField()) {
-            Word header = ObjectHeaderImpl.readHeaderFromObject(obj);
-            if (probability(SLOW_PATH_PROBABILITY, ObjectHeaderImpl.hasIdentityHashFromAddressInline(header))) {
-                addIdentityHashField = true;
-                copySize = LayoutEncoding.getSizeFromObjectInlineInGC(obj, true);
-            }
-        }
-
-        if (probability(VERY_SLOW_PATH_PROBABILITY, dest.isNull())) {
-            return WordFactory.zero();
-        }
-
-        /*
-         * This does a direct memory copy, without regard to whether the copied data contains object
-         * references. That's okay, because all references in the copy are visited and overwritten
-         * later on anyways (the card table is also updated at that point if necessary).
-         */
-        Pointer originalMemory = Word.objectToUntrackedPointer(obj);
-        Log.noopLog().string("Copying object from ").zhex(originalMemory).character('-').zhex(originalMemory.add(originalSize))
-                .string(" (").unsigned(originalSize).string(" B)")
-                .string(" to ").zhex(dest).character('-').zhex(dest.add(copySize))
-                .string(" (").unsigned(copySize).string(" B)")
-                .newline().flush();
-        UnmanagedMemoryUtil.copyLongsForward(originalMemory, dest, copySize);
-
-        if (probability(SLOW_PATH_PROBABILITY, addIdentityHashField)) {
-            Log.log().string("Added identity hash field!\n").flush();
-            Object copy = dest.toObject();
-            // Must do first: ensures correct object size below and in other places
-            int value = IdentityHashCodeSupport.computeHashCodeFromAddress(obj);
-            int offset = LayoutEncoding.getOptionalIdentityHashOffset(copy);
-            ObjectAccess.writeInt(copy, offset, value, IdentityHashCodeSupport.IDENTITY_HASHCODE_LOCATION);
-            ObjectHeaderImpl.getObjectHeaderImpl().setIdentityHashInField(copy);
-        }
-
-        return copySize;
-    }
-
-    /**
-     * Binary layout:
-     * <pre>
-     * 64-bit mode:
-     * +-------------------------+---------------+-----------------------+
-     * | relocation pointer (8B) | gap size (4B) | next plug offset (4B) |
-     * +-------------------------+---------------+-----------------------+
-     *                                                                   ^p
-     * 32-bit mode:
-     * +-------------------------+---------------+-----------------------+
-     * | relocation pointer (4B) | gap size (2B) | next plug offset (2B) |
-     * +-------------------------+---------------+-----------------------+
-     *                                                                   ^p
-     * </pre>
-     */
-    private static class RelocationInfo {
-
-        private static void writeRelocationPointer(Pointer p, Pointer relocationPointer) {
-            if (ObjectHeaderImpl.getReferenceSize() == Integer.BYTES) {
-                // TODO
-                assert false;
-            } else {
-                p.writeWord(-16, relocationPointer);
-            }
-        }
-
-        private static Pointer readRelocationPointer(Pointer p) {
-            if (ObjectHeaderImpl.getReferenceSize() == Integer.BYTES) {
-                // TODO
-                assert false;
-                return WordFactory.nullPointer();
-            } else {
-                return p.readWord(-16);
-            }
-        }
-
-        private static void writeGapSize(Pointer p, int gapSize) {
-            if (ObjectHeaderImpl.getReferenceSize() == Integer.BYTES) {
-                // TODO
-                assert false;
-            } else {
-                p.writeInt(-8, gapSize);
-            }
-        }
-
-        private static int readGapSize(Pointer p) {
-            if (ObjectHeaderImpl.getReferenceSize() == Integer.BYTES) {
-                // TODO
-                assert false;
-                return 0;
-            } else {
-                return p.readInt(-8);
-            }
-        }
-
-        private static void writeNextPlugOffset(Pointer p, int offset) {
-            if (ObjectHeaderImpl.getReferenceSize() == Integer.BYTES) {
-                // TODO
-                assert false;
-            } else {
-                p.writeInt(-4, offset);
-            }
-        }
-
-        private static int readNextPlugOffset(Pointer p) {
-            if (ObjectHeaderImpl.getReferenceSize() == Integer.BYTES) {
-                // TODO
-                assert false;
-                return 0;
-            } else {
-                return p.readInt(-4);
-            }
-        }
-
-        private static Pointer getNextRelocationInfo(Pointer p) {
-            int offset = readNextPlugOffset(p);
-            if (offset == 0) {
-                return WordFactory.nullPointer();
-            }
-            return p.add(offset);
-        }
-
-        /**
-         * Special implementation that is capable of handling gaps by reading the relocation info.
-         *
-         * @see HeapChunk#walkObjectsFrom
-         * @see AlignedHeapChunk#walkObjects
-         */
-        private static void walkObjects(AlignedHeapChunk.AlignedHeader chunkHeader, ObjectVisitor visitor) {
-            Pointer cursor = AlignedHeapChunk.getObjectsStart(chunkHeader);
-            Pointer top = HeapChunk.getTopPointer(chunkHeader); // top cannot move in this case
-            Pointer relocationInfo = chunkHeader.getFirstRelocationInfo();
-
-            while (cursor.belowThan(top)) {
-
-                // jump gaps
-                if (relocationInfo.isNonNull()) {
-                    int gapSize = RelocationInfo.readGapSize(relocationInfo);
-                    if (cursor.aboveOrEqual(relocationInfo.subtract(gapSize))) {
-                        cursor = relocationInfo;
-                        relocationInfo = RelocationInfo.getNextRelocationInfo(relocationInfo);
-                        continue;
-                    }
-                }
-
-                Object obj = cursor.toObject();
-                UnsignedWord objSize = LayoutEncoding.getSizeFromObjectInlineInGC(obj);
-
-                if (!visitor.visitObject(obj)) {
-                    return;
-                }
-
-                cursor = cursor.add(objSize);
-            }
-        }
-    }
-
-    static Object getRelocatedObject(Pointer p) {
-        return getRelocatedObjectPointer(p).toObject();
-    }
-
-    static Pointer getRelocatedObjectPointer(Pointer p) {
-        assert ObjectHeaderImpl.isAlignedObject(p.toObject()) : "Unaligned objects are not supported!";
-
-        AlignedHeapChunk.AlignedHeader aChunk = AlignedHeapChunk.getEnclosingChunkFromObjectPointer(p);
-        Pointer relocationInfo = aChunk.getFirstRelocationInfo();
-        if (relocationInfo.isNull() || p.belowThan(relocationInfo)) {
-            return p; // not relocated
-        }
-
-        Pointer nextRelocationInfo = RelocationInfo.getNextRelocationInfo(relocationInfo);
-        while (nextRelocationInfo.isNonNull() && nextRelocationInfo.belowOrEqual(p)) {
-            relocationInfo = nextRelocationInfo;
-            nextRelocationInfo = RelocationInfo.getNextRelocationInfo(relocationInfo);
-        }
-
-        assert relocationInfo.belowOrEqual(p);
-
-        Pointer relocationPointer = RelocationInfo.readRelocationPointer(relocationInfo);
-        Pointer relocationOffset = p.subtract(relocationInfo);
-
-        return relocationPointer.add(relocationOffset);
     }
 }
