+++ substratevm/src/com.oracle.svm.core.genscavenge/src/com/oracle/svm/core/genscavenge/UnalignedHeapChunk.java
@@ -71,9 +71,7 @@ import jdk.graal.compiler.word.Word;
  * </pre>
  *
  * The HeapChunk fields can be accessed as declared fields. The size of the card table depends on
- * the used {@link Rememb
- * eredSet} implementation and may even be zero.
- *
+ * the used {@link RememberedSet} implementation and may even be zero.
  * 
  * In this implementation, I am only implementing imprecise card remembered sets, so I only need one
  * entry for the whole Object. But for consistency I am treating it as a 1-element table.
